fix: remove unneccesary logic from fetchCharacter hook
